move piwik API wrapper functions to their own lib file
